Check constructor kwargs for 'None' before expanding them
When converting a JSON object in an API response to a python object the
generated code treats the dictionary from the converted JSON object as a
'kwargs' dictionary and passes the exploded dictionary to the
corresponding model's constructor. I.e.:

        constructor(**dict_from_json)

However if 'dict_from_json' is 'None' (converted from a 'null' JSON
value) this "explode" operator will fail (because None is not a
dictionary).

Therefore the generated code now first checks whether or not the
dictionary from JSON is 'None' and proceeds to explode the dictionary
only if it is NOT None. For cases where it is None it returns None
rather than call the constructor.

category: defect_fix
defect: #27